# infra/apigateway/openapi.yaml
swagger: "2.0"
info:
  title: nekkoflix-api
  description: nekkoflix Backend API Gateway
  version: "1.0.0"
schemes:
  - https
produces:
  - application/json
paths:
  /health:
    get:
      summary: Health check
      operationId: health
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/HealthResponse"
        "502":
          description: Backend error
  /generate:
    options:
      summary: CORS preflight for generate
      operationId: generateOptions
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "204":
          description: No Content
    post:
      summary: Generate a cat-tailored video
      operationId: generate
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 360.0
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/GenerateRequest"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/GenerateResponse"
        "400":
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Unauthorized
        "502":
          description: Upstream backend error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "504":
          description: Upstream timeout
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sessions/{session_id}/reaction-upload-url:
    options:
      summary: CORS preflight for reaction upload URL
      operationId: reactionUploadUrlOptions
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "204":
          description: No Content
    post:
      summary: Issue a signed upload URL for a reaction video
      operationId: issueReactionUploadUrl
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      parameters:
        - in: path
          name: session_id
          required: true
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ReactionUploadUrlResponse"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "409":
          description: Invalid session state
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sessions/{session_id}/reaction:
    options:
      summary: CORS preflight for reaction completion
      operationId: reactionCompleteOptions
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "204":
          description: No Content
    post:
      summary: Notify backend that the reaction video upload has completed
      operationId: registerReactionVideo
      x-google-backend:
        address: ${backend_url}
        path_translation: APPEND_PATH_TO_ADDRESS
        deadline: 60.0
      parameters:
        - in: path
          name: session_id
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ReactionUploadCompleteRequest"
      responses:
        "202":
          description: Accepted
          schema:
            $ref: "#/definitions/ReactionUploadResponse"
        "400":
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Unauthorized
        "502":
          description: Upstream backend error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  GenerateRequest:
    type: object
    required:
      - mode
      - image_base64
    properties:
      mode:
        type: string
        enum:
          - experience
          - production
      image_base64:
        type: string
      audio_base64:
        type: string
      user_context:
        type: string
  GenerateResponse:
    type: object
    required:
      - session_id
      - video_url
      - state_key
      - template_id
      - template_name
    properties:
      session_id:
        type: string
      video_url:
        type: string
      state_key:
        type: string
      template_id:
        type: string
      template_name:
        type: string
  HealthResponse:
    type: object
    required:
      - status
      - environment
    properties:
      status:
        type: string
        enum:
          - ok
      environment:
        type: string
  ReactionUploadUrlResponse:
    type: object
    required:
      - session_id
      - upload_url
      - reaction_video_gcs_uri
      - expires_in_seconds
    properties:
      session_id:
        type: string
      upload_url:
        type: string
      reaction_video_gcs_uri:
        type: string
      expires_in_seconds:
        type: integer
  ReactionUploadCompleteRequest:
    type: object
    required:
      - reaction_video_gcs_uri
    properties:
      reaction_video_gcs_uri:
        type: string
  ReactionUploadResponse:
    type: object
    required:
      - session_id
      - status
      - reaction_video_gcs_uri
    properties:
      session_id:
        type: string
      status:
        type: string
        enum:
          - accepted
      reaction_video_gcs_uri:
        type: string
  ErrorResponse:
    type: object
    required:
      - error_code
      - message
    properties:
      error_code:
        type: string
      message:
        type: string